feat (postgres): add CoinBuyHistory
add CoinBuyHistory, delete bank related tables and entities

+++ src/main/java/org/highload/model/stock/CoinPrice.java
@@ -17,5 +17,5 @@ public class CoinPrice {
     private Coin coin;
 
     @Column(name = "price_dollars")
-    private Integer priceInDollars;
+    private Long priceInDollars;
 }
